* develop: (812 commits)
  MueLu: comment out mmayr research subdirectory
  Enable Tacho in Kokkos promotion configuration
  Shylu/Tacho - fix the unit test failure described in #1935
  Tpetra: Fix spelling errors (see #1956)
  Intrepid2: implementing new design suggested by Kyungjoo to have work view of rank 1, instead of having a work view of rank 2 or 3.            In this case we have to expose Fad dimension.
  Intrepid2: introduce new function get_dimension_scalar(view) that returns the dimension_scalar(view),                for non pod view types (defined e.g. in Sacado) and returns 1 for pod types.
  HHG: 1D test code to figure out issues with parallelism
  MueLu tutorial: correct instructions in README file
  MueLu/HHG: setup row maps for each region
  MueLu/HHG: setup data structure to store all nodes in each region
  MueLu/HHG: read proc info from file
  MueLu/HHG: write region matrix information to file (using Matlab)
  MueLu/HHG: attempt to fix indices
  MueLu/HHG: First draft of regional splitting of matrices
  MueLu/HHG by Max: move creation of Xpetra::MatrixSplitting to driver
  MueLu/HHG by Max: move creation of RegionHandler to Test_xpetra.cpp
  Single versus double colon.
  Fixed FS example.
  ShyLU: Add missing TRIBITS_PACKAGE_DEF() calls (TriBITSPub/TriBITS#200)
  Enable ST code in ST extrab builds
  ...

+++ packages/teuchos/core/src/Teuchos_HashUtils.hpp
@@ -150,7 +150,7 @@ namespace Teuchos
        * but they are both true hashes anyway, so leaving it !
        * */
       const char* str = x.c_str();
-      int len = x.length();
+      int len = static_cast<int>(x.length());
       int step = len/4 + 1;
       int base = 1;
       int rtn = 0;
